Extract file content loading contracts partial
